requirements.txt
torch
lightning
hydra_colorlog
wandb
swanlab
deepspeed